refactor : 패키지 변경 및 테스트 분리
