+++ src/main/java/roomescape/admin/AdminController.java
@@ -1,4 +1,4 @@
-package roomescape;
+package roomescape.admin;
 
 import org.springframework.stereotype.Controller;
 import org.springframework.web.bind.annotation.GetMapping;
