+++ src/main/java/roomescape/domain/Times.java
@@ -1,4 +1,4 @@
-package roomescape.times;
+package roomescape.domain;
 
 import java.util.Objects;
