+++ src/main/java/roomescape/reservations/ReservationRepository.java
@@ -1,4 +1,4 @@
-package roomescape;
+package roomescape.reservations;
 
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.jdbc.core.JdbcTemplate;
@@ -6,6 +6,7 @@ import org.springframework.jdbc.core.namedparam.BeanPropertySqlParameterSource;
 import org.springframework.jdbc.core.namedparam.SqlParameterSource;
 import org.springframework.jdbc.core.simple.SimpleJdbcInsert;
 import org.springframework.stereotype.Repository;
+import roomescape.domain.Reservation;
 import roomescape.dto.RequestReservation;
 
 import java.util.List;
