+++ src/main/java/roomescape/times/TimesController.java
@@ -3,6 +3,7 @@ package roomescape.times;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Controller;
 import org.springframework.web.bind.annotation.*;
+import roomescape.domain.Times;
 import roomescape.dto.RequestTimes;
 import roomescape.dto.ResponseTimes;
