+++ src/test/java/roomescape/admin/ApiTest.java
@@ -1,4 +1,4 @@
-package roomescape;
+package roomescape.admin;
 
 import io.restassured.RestAssured;
 import org.junit.jupiter.api.BeforeEach;
@@ -7,12 +7,10 @@ import org.springframework.boot.test.context.SpringBootTest;
 import org.springframework.boot.test.web.server.LocalServerPort;
 import org.springframework.test.annotation.DirtiesContext;
 
-import static org.hamcrest.Matchers.is;
-
 
 @SpringBootTest(webEnvironment = SpringBootTest.WebEnvironment.RANDOM_PORT)
 @DirtiesContext(classMode = DirtiesContext.ClassMode.BEFORE_CLASS)
-public class MissionStepTest {
+public class ApiTest {
 
     @LocalServerPort
     int port;
@@ -36,14 +34,5 @@ public class MissionStepTest {
                 .when().get("/admin/reservation")
                 .then().log().all()
                 .statusCode(200);
-    }
-
-    @Test
-    void 예약_목록들을_가져온다() {
-        RestAssured.given().log().all()
-                .when().get("/reservations")
-                .then().log().all()
-                .statusCode(200)
-                .body("size()", is(0)); // 아직 생성 요청이 없으니 Controller에서 임의로 넣어준 Reservation 갯수 만큼 검증하거나 0개임을 확인하세요.
     }
 }
