+++ src/test/java/roomescape/reservations/ApiTest.java
@@ -1,4 +1,4 @@
-package roomescape;
+package roomescape.reservations;
 
 import io.restassured.RestAssured;
 import io.restassured.http.ContentType;
@@ -15,7 +15,7 @@ import static org.hamcrest.Matchers.is;
 
 @SpringBootTest(webEnvironment = SpringBootTest.WebEnvironment.RANDOM_PORT)
 @DirtiesContext(classMode = DirtiesContext.ClassMode.BEFORE_EACH_TEST_METHOD)
-public class CrudApiTest {
+public class ApiTest {
 
     @LocalServerPort
     int port;
@@ -23,6 +23,15 @@ public class CrudApiTest {
     @BeforeEach
     void beforeEach() {
         RestAssured.port = port;
+    }
+
+    @Test
+    void 예약_목록들을_가져온다() {
+        RestAssured.given().log().all()
+                .when().get("/reservations")
+                .then().log().all()
+                .statusCode(200)
+                .body("size()", is(0)); // 아직 생성 요청이 없으니 Controller에서 임의로 넣어준 Reservation 갯수 만큼 검증하거나 0개임을 확인하세요.
     }
 
     @Test
